Refactor classes. Fix load bug.
