+++ src/linearInterpolation/userInterface/mainFrame/userInput/inner/ValuesPanel.java
@@ -1,4 +1,4 @@
-package linearInterpolation.userInterface.panels;
+package linearInterpolation.userInterface.mainFrame.userInput.inner;
 
 import linearInterpolation.model.Interpolation;
 
@@ -9,53 +9,21 @@ import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.Collection;
 
-import static javax.swing.JScrollPane.HORIZONTAL_SCROLLBAR_NEVER;
-import static javax.swing.JScrollPane.VERTICAL_SCROLLBAR_ALWAYS;
+import static javax.swing.ScrollPaneConstants.HORIZONTAL_SCROLLBAR_NEVER;
+import static javax.swing.ScrollPaneConstants.VERTICAL_SCROLLBAR_ALWAYS;
 
-public class InitialValuesPanel extends JPanel {
-    public final int defaultValuesCount = 5;
+// TODO: initialize interpolation
+public class ValuesPanel extends JPanel {
     private final NumberFormat numberFormat = NumberFormat.getNumberInstance();
     private final NumberFormat intFormat = NumberFormat.getIntegerInstance();
 
+    private Interpolation interpolation;
     private ArrayList<JTextField> xValuesFields;
     private ArrayList<JTextField> yValuesFields;
 
-    private Interpolation interpolation;
-    private JPanel initializingPanel;
-    private JPanel valuesPanel;
-    private JFormattedTextField valuesCountField;
-
-    public InitialValuesPanel(Interpolation interpolation) {
-        this.interpolation = interpolation;
-        setBorder(BorderFactory.createTitledBorder("Input values"));
-        setLayout(new GridLayout(1, 2));
-        initializingPanel = new JPanel();
-        initializingPanel.setLayout(new BoxLayout(initializingPanel,
-                BoxLayout.Y_AXIS));
-        initializingPanel.add(createCountPanel());
-        valuesPanel = createValuesPanel(defaultValuesCount);
-        initializingPanel.add(valuesPanel);
-        add(initializingPanel);
-        add(createNewPointsPanel());
-    }
-
-    private JPanel createCountPanel() {
-        JPanel borderPanel = new JPanel(new BorderLayout());
-        JPanel countPanel = new JPanel(new FlowLayout());
-        JButton setButton = new JButton("Set values count");
-        valuesCountField = new JFormattedTextField(intFormat);
-
-        valuesCountField.setValue(5);
-        valuesCountField.setColumns(5);
-        setButton.addActionListener(e -> updateValuesPanel());
-        countPanel.add(valuesCountField);
-        countPanel.add(setButton);
-        borderPanel.add(countPanel, BorderLayout.WEST);
-        return borderPanel;
-    }
-
-    private JPanel createValuesPanel(int valuesCount) {
-        JPanel borderPanel = new JPanel(new BorderLayout());
+    public ValuesPanel(int valuesCount, Interpolation interpolation) {
+        setInterpolation(interpolation);
+        setLayout(new BorderLayout());
         JPanel valuesPanel = new JPanel(new BorderLayout());
         JPanel fieldsPanel = new JPanel(new GridLayout(valuesCount + 1, 2));
         JScrollPane scrollPane = new JScrollPane(fieldsPanel,
@@ -89,43 +57,13 @@ public class InitialValuesPanel extends JPanel {
         initializeButton.addActionListener(e -> initializeInterpolation());
         valuesPanel.add(scrollPane, BorderLayout.CENTER);
         valuesPanel.add(initializeButton, BorderLayout.SOUTH);
-        borderPanel.add(valuesPanel, BorderLayout.WEST);
-        return borderPanel;
+        add(valuesPanel, BorderLayout.WEST);
     }
 
-    private JPanel createNewPointsPanel() {
-        JPanel newPointsPanel = new JPanel();
-        JFormattedTextField xField = new JFormattedTextField(numberFormat);
-        xField.setColumns(5);
-        JButton addButton = new JButton("Add");
-        addButton.addActionListener(e -> {
-            try {
-                interpolation.addPoint(parseField(xField));
-            } catch (ParseException parseException) {
-                showNumberFormatErrorDialog();
-            }
-        });
-        newPointsPanel.setBorder(BorderFactory.createTitledBorder("Enter X"));
-        newPointsPanel.add(xField);
-        newPointsPanel.add(addButton);
-        return newPointsPanel;
+    public void setInterpolation(Interpolation interpolation) {
+        this.interpolation = interpolation;
     }
 
-    private void updateValuesPanel() {
-        int count;
-        try {
-            count = intFormat.parse(valuesCountField.getText()).intValue();
-        } catch (ParseException e) {
-            showNumberFormatErrorDialog();
-            return;
-        }
-        initializingPanel.remove(valuesPanel);
-        valuesPanel = createValuesPanel(count);
-        initializingPanel.add(valuesPanel);
-        valuesPanel.revalidate();
-    }
-
-    // TODO: finish the method (calculation button)
     private void initializeInterpolation() {
         try {
             Collection<Double> timeStamps = parseFields(xValuesFields);
@@ -157,9 +95,5 @@ public class InitialValuesPanel extends JPanel {
                 "Error. Wrong number format.",
                 "Error",
                 JOptionPane.ERROR_MESSAGE);
-    }
-
-    public void setInterpolation(Interpolation interpolation) {
-        this.interpolation = interpolation;
     }
 }
